+++ package.json
@@ -15,6 +15,7 @@
     "prepare": "husky"
   },
   "dependencies": {
+    "axios": "^1.10.0",
     "better-auth": "1.2.9-beta.9",
     "dotenv": "^16.5.0",
     "drizzle-orm": "^0.44.1",
