+++ src/lib/db/index.ts
@@ -1,2 +1,3 @@
 export * from "./instance";
 export * from "./schema";
+export * from "./types";
